Resolved issue when writing to HOSTS
Add-Content, Clear-Content, and Set-Content all needed to establish a file lock in order to write to the file. This caused errors to be thrown when enabling and disabling sites which often left the system is a partial state. Either some of the sites would be enabled or disabled.

I changed the code to use Out-File as it does not require a lock on the file. In addition I removed the last error check for an empty HOSTS file, because in theory that could be a valid case.