Addresses #9, now able to execute PowerShell scripts using a Windows Task
1. Get-DateTime needed to add -UseBasicParsing because the System account does not have Internet Explorer pre-configured. Using it without will cause the script to throw an error stating that the `internet explorer engine is not available`.
2. I need to task to fire at `LogOn` rather then at `StartUp` so that we can ensure it is running when the computer is being used.
3. I discovered a flaw in New-ScheduledTaskAction where incorrect characters were being added to the end of the argument string for the action. These characters replaced the final double quote and caused the PowerShell scripts to fail to execute. These characters were found by leveraging NotePad++ to change the encoding of the text. I filed a bug with Microsoft but it appears that not much attention is being paid on this site. Luckily I happened to find a work around by concatenating a literally typed double quote to the end of the string.

Still to do, I want to update the installer script to start the scheduled task instead of calling Keep Alive. Keep Alive is designed to not end, so it keeps the installer script running even after installation is complete. Also if that were to be closed at any point, the task is not actually running until the next log on. So it would be possible for the sites, and applications in the future, to become re-enabled without the monitoring in place. Start the task should resolve that gap.